+++ src/php/Lint/Infrastructure/Command/LintCommand.php
@@ -6,6 +6,7 @@ namespace Phel\Lint\Infrastructure\Command;
 
 use Gacela\Framework\ServiceResolver\ServiceMap;
 use Gacela\Framework\ServiceResolverAwareTrait;
+use Phel\Filesystem\Application\PhelProjectDirectory;
 use Phel\Lint\Application\Cache\LintCache;
 use Phel\Lint\Application\Config\RuleSettings;
 use Phel\Lint\Application\Formatter\HumanFormatter;
@@ -194,6 +195,7 @@ final class LintCommand extends Command
             return null;
         }
 
+        PhelProjectDirectory::ensure($cwd);
         $cacheDir = rtrim($cwd, '/') . '/' . LintConfig::defaultCacheDir();
 
         return $this->getFacade()->createCache($cacheDir);